backend/pyproject.toml
[tool.poetry]
name = "dailyinsightai"
version = "0.1.0"
description = "This project aims to combine Python, MongoDB, AI to build a journaling/idea management app"
authors = ["Nicoleta Mocanu <nicoleta.mocanu87@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11"
openai = ">=1.61.1,<2.0.0"
pymongo = ">=4.11,<5.0"
pandas = ">=2.2.3,<3.0.0"
flask = ">=3.1.0,<4.0.0"
fastapi = ">=0.115.8,<0.116.0"
uvicorn = ">=0.34.0,<0.35.0"
python-dotenv = ">=1.0.1,<2.0.0"
ruff = "^0.11.7"
black = "^25.1.0"
h11 = ">=0.16.0"
Jinja2 = ">=3.1.6"
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

